+++ raiden-dapp/src/components/AmountInput.vue
@@ -33,13 +33,13 @@ export default class AmountInput extends Vue {
   label?: string;
   @Prop({})
   disabled!: boolean;
-  @Prop({ default: '0.00' })
+  @Prop({ required: true })
   value!: string;
   @Prop()
   token?: Token;
   @Prop({ default: false, type: Boolean })
   limit!: boolean;
-  @Prop({ default: '0.0', type: String })
+  @Prop({ default: '', type: String })
   placeholder!: string;
   @Prop({ required: false, default: () => Zero })
   max!: BigNumber;
@@ -73,7 +73,7 @@ export default class AmountInput extends Vue {
       !this.limit ||
       (v && this.hasEnoughBalance(v, this.max)) ||
       this.$parent.$t('amount-input.error.not-enough-funds', {
-        funds: BalanceUtils.toUnits(this.max, this.token!.decimals || 18),
+        funds: BalanceUtils.toUnits(this.max, this.token!.decimals ?? 18),
         symbol: this.token!.symbol
       })
   ];
@@ -81,7 +81,7 @@ export default class AmountInput extends Vue {
   private noDecimalOverflow(v: string) {
     return (
       AmountInput.numericRegex.test(v) &&
-      !BalanceUtils.decimalsOverflow(v, this.token!.decimals || 18)
+      !BalanceUtils.decimalsOverflow(v, this.token!.decimals ?? 18)
     );
   }