+++ raiden-dapp/src/filters.ts
@@ -31,21 +31,25 @@ export default class Filters {
   }
 
   static displayFormat(amount: BigNumber, decimals?: number): string {
-    const units = BalanceUtils.toUnits(amount, decimals || 18);
+    const units = BalanceUtils.toUnits(amount, decimals ?? 18);
     const deposit = parseFloat(units);
     if (deposit === 0) {
-      return '0.0';
+      return decimals === 0 ? '0' : '0.0';
     } else if (deposit < 0.000001) {
       return '<0.000001';
     } else {
-      const [integerPart, decimalPart] = split(units, '.');
+      return Filters.formatValue(units, decimals);
+    }
+  }
 
-      if (decimalPart && decimalPart.length > 6) {
-        let newDecimal = decimalPart.substring(0, 6);
-        return `≈${integerPart}.${newDecimal}`;
-      } else {
-        return units;
-      }
+  private static formatValue(units: string, decimals: number | undefined) {
+    const [integerPart, decimalPart] = split(units, '.');
+
+    if (decimalPart && decimalPart.length > 6) {
+      let newDecimal = decimalPart.substring(0, 6);
+      return `≈${integerPart}.${newDecimal}`;
+    } else {
+      return decimals === 0 ? integerPart : units;
     }
   }
 
@@ -54,7 +58,7 @@ export default class Filters {
   }
 
   static toUnits = (wei: BigNumber, decimals?: number) =>
-    BalanceUtils.toUnits(wei, decimals || 18);
+    BalanceUtils.toUnits(wei, decimals ?? 18);
 
   static formatDate = (value: Date): string => {
     return `${new Intl.DateTimeFormat('en-US').format(